Improve consistency of Cardano module with others
